+++ lib/msal-browser/test/config/Configuration.spec.ts
@@ -1,6 +1,6 @@
 import { Configuration, buildConfiguration, DEFAULT_POPUP_TIMEOUT_MS, DEFAULT_IFRAME_TIMEOUT_MS } from "../../src/config/Configuration";
 import { TEST_CONFIG, TEST_URIS } from "../utils/StringConstants";
-import { LogLevel, Constants, ClientConfigurationError, ClientConfigurationErrorMessage, AzureCloudInstance } from "@azure/msal-common";
+import { LogLevel, Constants, AzureCloudInstance } from "@azure/msal-common";
 import sinon from "sinon";
 import { BrowserCacheLocation } from "../../src/utils/BrowserConstants";